+++ src/components/common/TextEditor/utils/handlers/getHtmlHeadings.ts
@@ -1,18 +1,23 @@
-'use client';
+import * as cheerio from 'cheerio';
+
 export interface HtmlHeadingItems {
   level: number;
   text: string;
   id: string | null;
 }
 
-export const getHtmlHeadings = (html: string) => {
-  const parser = new DOMParser();
-  const doc = parser.parseFromString(html, 'text/html');
-  const headings = Array.from(doc.querySelectorAll('h1, h2, h3, h4, h5, h6'));
+export const getHtmlHeadings = (html: string): HtmlHeadingItems[] => {
+  const $ = cheerio.load(html);
+  const headings: HtmlHeadingItems[] = [];
 
-  return headings.map((el) => ({
-    level: parseInt(el.tagName[1]),
-    text: el.textContent?.trim() ?? '',
-    id: el.id || null,
-  }));
+  $('h1, h2, h3, h4, h5, h6').each((_, elem) => {
+    const tagName = elem.tagName.toLowerCase(); // ex) 'h2'
+    const level = parseInt(tagName[1]); // '2' → 2
+    const text = $(elem).text().trim();
+    const id = $(elem).attr('id') || null;
+
+    headings.push({ level, text, id });
+  });
+
+  return headings;
 };
